# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "健身计划管理系统API",
        "title": "Flow Link Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/stats/calendar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户指定月份的训练日历数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "获取日历数据",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 2024,
                        "description": "年份",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "月份",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.CalendarDay"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/muscle-groups": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的肌群训练统计数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "获取肌群训练统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "month",
                        "description": "统计周期(week/month/year)",
                        "name": "period",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MuscleGroupStats"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/personal-records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的个人最佳记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "获取个人记录",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.PersonalRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/training": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取用户的训练统计数据，支持按周期查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "统计"
                ],
                "summary": "获取训练统计",
                "parameters": [
                    {
                        "type": "string",
                        "default": "week",
                        "description": "统计周期(week/month/year)",
                        "name": "period",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "开始日期",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "结束日期",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/domain.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.TrainingStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权访问",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.CalendarDay": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "日期 YYYY-MM-DD",
                    "type": "string"
                },
                "hasTraining": {
                    "description": "是否有训练",
                    "type": "boolean"
                },
                "totalDuration": {
                    "description": "总时长",
                    "type": "integer"
                },
                "trainingCount": {
                    "description": "训练次数",
                    "type": "integer"
                }
            }
        },
...
        "domain.DailyStats": {
            "type": "object",
            "properties": {
                "calories": {
                    "description": "当日消耗卡路里",
                    "type": "integer"
                },
                "date": {
                    "description": "日期 YYYY-MM-DD",
                    "type": "string"
                },
                "duration": {
                    "description": "当日训练时长",
                    "type": "integer"
                },
                "sets": {
                    "description": "当日总组数",
                    "type": "integer"
                },
                "trainingCount": {
                    "description": "当日训练次数",
                    "type": "integer"
                },
                "weight": {
                    "description": "当日总重量",
                    "type": "number"
                }
            }
        },
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.MuscleGroupStats": {
            "type": "object",
            "properties": {
                "muscleGroup": {
                    "description": "肌群名称",
                    "type": "string"
                },
                "percentage": {
                    "description": "占比百分比",
                    "type": "integer"
                },
                "totalWeight": {
                    "description": "总重量",
                    "type": "number"
                },
                "trainingCount": {
                    "description": "训练次数",
                    "type": "integer"
                }
            }
        },
...
        "domain.PersonalRecord": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "创建日期",
                    "type": "string"
                },
                "exerciseName": {
                    "description": "训练项目名称",
                    "type": "string"
                },
                "maxWeight": {
                    "description": "最大重量",
                    "type": "number"
                },
                "recordId": {
                    "description": "记录ID",
                    "type": "integer"
                }
            }
        },
...
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "domain.TrainingStats": {
            "type": "object",
            "properties": {
                "avgDuration": {
                    "description": "平均训练时长",
                    "type": "integer"
                },
                "avgWeight": {
                    "description": "平均单次重量",
                    "type": "number"
                },
                "dailyStats": {
                    "description": "每日统计数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.DailyStats"
                    }
                },
                "endDate": {
                    "description": "统计结束日期",
                    "type": "string"
                },
                "favoriteExercise": {
                    "description": "最常做的训练项目",
                    "type": "string"
                },
                "mostTrainedMuscle": {
                    "description": "训练最多的肌群",
                    "type": "string"
                },
                "period": {
                    "description": "统计周期(week/month/year)",
                    "type": "string"
                },
                "startDate": {
                    "description": "统计开始日期",
                    "type": "string"
                },
                "totalCalories": {
                    "description": "总消耗卡路里",
                    "type": "integer"
                },
                "totalDuration": {
                    "description": "总训练时长(分钟)",
                    "type": "integer"
                },
                "totalSets": {
                    "description": "总组数",
                    "type": "integer"
                },
                "totalTrainingCount": {
                    "description": "总训练次数",
                    "type": "integer"
                },
                "totalWeight": {
                    "description": "总重量(kg)",
                    "type": "number"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}